게임함수 파일 분리
