+++ main.py
@@ -1,71 +1,5 @@
-##기억력테스트 게임(시장에 가면 게임)
-import random
-import time
-sequence = []  # 전체 아이템 시퀀스
-item_pool = ["사과", "배", "수박", "책", "모자", "바지", "치약", "우산", "신발", "귤"]
-#현재까지 물건 기억
-def print_sequence():
-    print("\n현재까지의 물건 목록:")
-    for i, item in enumerate(sequence):
-        print(f"{i+1}. {item}")
-    print()
-#유저 차례
-def user_turn():
-    print_sequence()
-    print("지금까지의 물건을 순서대로 입력하세요.")
-    for i, item in enumerate(sequence):
-        answer = input(f"{i+1}번째 물건: ")
-        if answer.strip() != item:
-            print("틀렸습니다! 게임 오버.")
-            return False
-    # 새로운 아이템 추가
-    while True:
-        new_item = input("새로운 물건을 입력하세요: ").strip()
-        if new_item and new_item not in sequence:
-            sequence.append(new_item)
-            break
-        else:
-            print("이미 입력된 물건이거나 잘못된 입력입니다.")
-    return True
-#컴퓨터 차례(미리 정의 된 물건들 이름 다쓰면 out)
-def cpu_turn(cpu_name):
-    print(f"{cpu_name}의 차례입니다...")
-    time.sleep(1)
-    # 기존 시퀀스 암기 시뮬레이션
-    for i, item in enumerate(sequence):
-        time.sleep(0.5)
-        print(f"{cpu_name}: {item}")
-    # 새로운 아이템 추가
-    remaining_items = list(set(item_pool) - set(sequence))
-    if not remaining_items:
-        print("아이템 풀이 모두 소진되었습니다!")
-        return False
-    new_item = random.choice(remaining_items)
-    print(f"{cpu_name}: {new_item}")
-    sequence.append(new_item)
-    return True
-#게임 실행함수
-def gamegichan(players, user_name):
-    print("시장에 가면~ 🎵")
-    round_num = 1
-    sequence.clear()  # 이전 게임의 아이템 시퀀스 초기화
-    while True:
-        print(f"\n===== Round {round_num} =====")
-        for player in players:
-            print(f"\n{player['name']}의 차례입니다.")
-            if player['name'] == user_name:
-                if not user_turn():
-                    print(f"{player['name']}님이 틀렸습니다!")
-                    return
-            else:
-                if not cpu_turn(player['name']):
-                    print(f"{player['name']}의 차례에서 아이템 풀이 소진되었습니다!")
-                    return
-        round_num += 1
-# 김동근
-#김동수
-#김지은
-#박주은
+from gichan import *
+
 #플레이어 설정
 #동수,지은,기찬,주은,동근에 대한 이름과 치사량 저장
 friends_pool = [
